fix: apply runtime settings immediately in editor
