+++ SherpaONNXUnity/Packages/com.eitan.sherpa-onnx-unity/Runtime/Core/Integration/SherpaONNXRuntimeSettingsApplier.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 226d00f0fa8e4ee4b4f9b2b57b7e10bf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
